Add proposed definition of the GEDZIP media type
Since it was decided that the GEDZIP version and gedcom.ged file version
must match, then the version parameter in the media type can be used to
specify how to interpret the GEDZIP format.

But since .gdz files come without a media type, we still need to specify
how to interpret a .gdz file when you don't know if it's 7.0 or some
future version.

Fixes #5

Signed-off-by: Dave Thaler <dthaler@armidalesoftware.com>