Massive reordering of metasyntax blocks
Issue #172 recommended alphabetizing structures within metasyntax blocks. Doing that often broke up semantically-related structures, so I tried to document excepsions. The end result was a seven-part ordering heuristic (documented in the changed text itself) and I'm still not entirely happy with all of how it turned out. There was something familiar with `NOTE_STRUCTURE` always being near the end and so on...

However, the clarifying text that order doesn't matter except for the `Dataset :=` rule should definitely stay. And the reordering of alternatives in a choice production was a lot less tricky than the reordering of substructures, so that should probably happen even if the rest of this is rejected.