Possible fix for media types
An open question is whether to allow only registered types/subtypes with
FORM or also extension types.

The g7:FORM text under section 3 only references BCP 13, and use of that reference
which implies they must be registered, not extension types.  However
section 2 implied it could be either one, which is a contradiction.
This PR does not yet address that contradiction.

Signed-off-by: Dave Thaler <dthaler@armidalesoftware.com>