Add NO NAME
This involves several parts. Because `NO NAME` and `NO IMMI` would appear in the same context, they must have the same structure type. That means the name "`NON_EVENT_STRUCTURE`" is not ideal anymore (this changes it to `NEGATIVE_ASSERTION`) and that the enumeration set should be renamed from `enumset-EVEN` to `enumset-NO`.

This also patches a gap in the previous spec whereby `NO BIRT` could be a substructure of `FAM`, but does so with a "should" not a "must" to avoid breaking backwards compatibility.