Conversation draft of what a KIND might look like
See #315 for more on this idea and why it's KIND instead of TYPE.
Each `KIND` has its own structure type so that it can have its own enumeration set.

This draft uses `KIND <Enum> {0:1}`, but we could allow several KINDs,
either by `KIND <List:Enum> {0:1}` or `KIND <Enum> {0:M}`.
I think single-kind is simpler, but because of that didn't include the `GROUP` kind proposed in <https://github.com/FamilySearch/GEDCOM/issues/315#issuecomment-1564557254> as it would presumably usually be paried with another kind.

I found deciding what was and wasn't a kind of burrial or marriage to be tricky. For example,

- I included sky burrials, even though those are sometimes classes as a body preparation step like embalming instead,
but didn't include embalming.
- I didn't include cohabitating/partnership/companionship marriage kinds because I thought some researchers might class those as being closer to a `NO MARR` than a `MARR`, but I'm not sure that was the right decision.
- I found two different definitions of common-law marriage, so I split those into `COMMON_LAW` and `INFORMAL`, but I'm not confident that was the right split

Before discussing those topics, though, I hope this PR can facilitate discussion of the following questions:

1. If we add KIND in 7.1, should it be
    1. `{0:1}` or
    2. `{0:M}` or
    3. vary by event type (e.g. marriage is `{0:M}` but burrial is `{0:1}`)?
2. Is the "use TYPE instead of PHRASE" the right decision? Is it worded clearly enough? Is it OK to use `OTHER` with `PHRASE` recommended most places but `TYPE` recommended for `KIND OTHER`?
3. If we like this idea, how should we decide which events get KINDs?