Clarify that there are no relcoated standard records
This edit explicitly calls out a case that was forbidden already, but only implicitly so.
We had defined *relocated standard structures* in terms of their superstructures,
which records do not have
and had a blanket statement "all other non-standard structures are prohibited"
but had not expllicitly called out that included in those "other non-standard structures"
were relocating a standard record to a substructure or vice-versa.

Resolves #365