Clarify nickname
Add additional clarification to nickname, explaining the word's meaning in English which is not shared by several European countries. See [this comment](https://github.com/FamilySearch/GEDCOM/issues/473#issuecomment-2168463707) and the rest of issue #473 for more on why this clarification is needed.

Although #473's discussion covers many more topics, if we go back to the title and first question in the issue I think this resolves #473.