Define "principle date" (#492)
* Define "principle date"

As pointed out in #488 and #490, the definition of DATE includes the vague phrase "principle date" which could use some clarification. This is my effort to provide that clarification.

Note, if competing definitions of the principle date in these contexts exists then this suggestion could be seen as backwards-incompatible and may need to be reworded as a non-normative recommendation or note. That said, I'm not aware of any conflicting definitions.

Resolves #490

* typos

* typo