Conversation draft of adding place records to 7.1
This puts PLAC and SPLAC side-by-side, like NOTE and SNOTE. That is not the only possibility, and alternative PRs with other designs are anticipated to allow better comparison of the options.

This only addresses the record-vs-substucture topic. The additional substuctures that have also been considered for PLAC/_LOC/etc are prepared for by creating a `<<PLACE_DETIALS>>` production to be the home for such additions in a future PR.