Draft v8.0 replacement structure for personal names.
This draft was created based on conversation with members of the names working group in https://github.com/fisharebest/gedcom-name/pull/27 and #473. The text is mostly new, though, and may have failed to capture some elements of those conversations.

A separate v7.1 draft is anticipated once conversation on this draft stabilizes.