Clarify FilePath syntax regarding non-ASCII characters (#651)
* Clarify FilePath syntax regarding non-ASCII characters

Fixes #648

Defines the payload as just being a "valid URL string" (not "or URI
reference") since the former includes the latter.

Explicitly stated that URLs can contain non-ascii characters as
permitted in a "valid URL string".

Moved RFC 3986 link down to the bullet that already constrained local
(but not remote) file paths to be URI references.

Explicitly stated that a local path requires percent-encoding
non-ASCII characters, since the spec already required that to be the
case.

Signed-off-by: Dave Thaler <dthaler@armidalesoftware.com>

* Address PR feedback

Signed-off-by: Dave Thaler <dthaler@armidalesoftware.com>

---------

Signed-off-by: Dave Thaler <dthaler@armidalesoftware.com>
Co-authored-by: Dave Thaler <dthaler@armidalesoftware.com>