Add RESN to all records that did not already have it (#662)
* Add RESN to all records that did not already have it

Resolve #221

* Change some URIs to 7.1

Also add tooling to several build python programs to support a mix of 7.0 and 7.1

The specification says ([section 1.2](https://gedcom.io/specifications/FamilySearchGEDCOMv7.html#structures) paragraph 3) that a URI identifies a structure's meaning, payload type (including enumeration set), substructures, and the cardinality of substructures. It does not say it defines superstructures; thus we keep g7:RESN but change g7:record-RESN to g7.1:record-RESN.

That said, the YAML files have now lost that the old `g7:record-REPO` is a valid superstructure of `g7:NAME` as well as the new `g7.1:record-REPO` being one. That information cannot be extracted from the 7.1 spec alone; it would require merging the 7.0 and 7.1 spec's information.