Draft v8.0 TEMPLATE, STICKY, FLEX, ASSET, GROUP, PROOF, alongside V7
This PR models new structures for GEDCOM V8, to try and take that to a higher level.
It is based on TEMPLATEs and STICKYs. They describe the evidence in documents, on gravestones, in a census, and any other evidence "entity" that can hold genealogical information.
It has 3 huge enumtables so it becomes a bit "Data driven".
It is structured alongside GEDCOM 7.x (for now) to help users and developers get used to this new system. But before implementation the group might decide to totally disconnect it from V7.x and start a new direction.
In that case changes have to be applied to make this draft work.

While designing, time and effort have been spend to make it extendable, reliable, easily changeable, logical, solid and easy for users and developers.

It needs polishing, refinements, help from the group to find missing ends.
It has a huge extra file with examples, they are not there for fun (although some have some fun inside) but they are there as a testbed.
I heavily used that to make sure things would fit.

But I now want to show it all, I can hardly wait.😁